🌊 Feat: 하트 연결 완료
